refactor: move `compute_lagrange_poly` to `utils`
